Address code-review nits: dead branch and test placement
- plain-scan status_code.py: remove the unreachable `status_code is None`
  branch. fetch() always returns a Response (it raises on failure, and
  scan() skips this audit when the fetch failed), so the guard was dead
  code. Keep a comment warning against re-introducing the `if response`
  bug.
- plain-tunnel: move the private `_is_streaming_response` tests to
  tests/internal/ (they assert on a private method — a change-detector,
  not a public contract) and keep the URL-derivation tests in public/.
  Share the make_client helper via tests/conftest.py.
